Refactor podcast cover
