+++ modules/features/endofyear/src/main/java/au/com/shiftyjelly/pocketcasts/endofyear/components/PodcastCover.kt
@@ -1,4 +1,4 @@
-package au.com.shiftyjelly.pocketcasts.endofyear.util
+package au.com.shiftyjelly.pocketcasts.endofyear.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -7,33 +7,28 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Card
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.drawWithContent
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Matrix
+import androidx.compose.ui.graphics.drawscope.withTransform
+import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import au.com.shiftyjelly.pocketcasts.compose.components.PodcastImage
 
-@Composable
-fun PodcastCoverSmall(
-    uuid: String,
-    coverWidth: Dp,
-    modifier: Modifier = Modifier,
-) {
-    PodcastImage(
-        uuid = uuid,
-        cornerSize = 4.dp,
-        modifier = modifier.size(coverWidth)
-    )
-}
+private const val PodcastCoverRotationAngle = -30f
+private const val PodcastCoverSkew = 0.45f
 
 @Composable
-fun PodcastCoverBig(
+fun PodcastCover(
     uuid: String,
     coverWidth: Dp,
+    coverType: PodcastCoverType,
     modifier: Modifier = Modifier,
 ) {
     PodcastImage(
         uuid = uuid,
-        cornerSize = 8.dp,
+        cornerSize = if (coverType == PodcastCoverType.SMALL) 4.dp else 8.dp,
         modifier = modifier.size(coverWidth)
     )
 }
@@ -59,4 +54,22 @@ fun RectangleCover(
             modifier = modifier.fillMaxSize()
         ) {}
     }
+}
+
+fun Modifier.transformPodcastCover() =
+    graphicsLayer(rotationZ = PodcastCoverRotationAngle)
+        .drawWithContent {
+            withTransform({
+                val transformMatrix = Matrix()
+                transformMatrix.values[Matrix.SkewX] = PodcastCoverSkew
+                transform(transformMatrix)
+            }
+            ) {
+                this@drawWithContent.drawContent()
+            }
+        }
+
+enum class PodcastCoverType {
+    SMALL,
+    BIG
 }
